stbt.Keyboard: Update docstrings, including worked example
The example code is now all together at the top, instead of fragments of
code in each of the docstrings for `add_grid`, `enter_text`, etc.

To keep it brief I haven't shown an example with modes; once we write a
blog post / tutorial we can link to that (but we'd still leave this
basic example in the docs).